requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.2.0
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.1.7
django-cors-headers==4.7.0
django-shortuuidfield==0.1.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
mysqlclient==2.2.7
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
pytz==2025.1
service-identity==24.2.0
setuptools==77.0.1
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
twisted-iocpsupport==25.2.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
zope.interface==7.2
